Certificate project server

+++ client/src/pages/TaskList.jsx
@@ -1,10 +1,12 @@
-import { redirect, useNavigate, useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import Button from "../components/Button";
 import Input from "../components/Input";
 import { useDispatch, useSelector } from "react-redux";
 import { addTask, addTaskList, deleteTask, deleteTaskList } from "../reducers/taskManagerReducer";
-import { useState } from "react";
-import TaskDetail from "./TaskDetails";
+import { useEffect, useState } from "react";
+import TaskDetail from "../components/TaskDetails";
+import tasklistService from "../services/tasklist-service";
+import taskService from "../services/task-service";
 
 const taskListInitialData = {
     id: '',
@@ -21,9 +23,11 @@ const taskInitialData = {
 };
 
 const TaskList = () => {
-    const updateTaskList = (state) => state.taskManager.taskListIndex.find(item => item.id == id)
+    const updateTaskList = (state) => state.taskManager.taskListIndex?.find(item => item.id == id)
     const { id } = useParams();
     const taskList = useSelector(updateTaskList);
+    const user = useSelector(state => state.taskManager.user);
+    const [error, setError] = useState();
 
     const dispatch = useDispatch();
     const navigate = useNavigate();
@@ -32,8 +36,17 @@ const TaskList = () => {
     const [errorTask, setErrorTask] = useState(null);
     const [errorList, setErrorList] = useState(null);
 
+    
+    useEffect(() => {
+        setData(taskList || taskListInitialData)
+    }, [taskList]);
+
     function handleGoBack() {
         navigate('/');
+    }
+
+    const setErrorMessage = (error) => {
+        setError(error?.response?.data?.error || 'Error to execute the opperation.');
     }
 
     function handleFieldChange(e) {
@@ -59,7 +72,7 @@ const TaskList = () => {
 
     const reloadData = (taskListId) => {
         handleGoBack();
-        setTimeout(()=> navigate(`/task-list/${taskListId}`));
+        setTimeout(() => navigate(`/task-list/${taskListId}`), 1);
     }
 
     const handleClickSave = (e) => {
@@ -71,32 +84,44 @@ const TaskList = () => {
             return;
         }
         setErrorList(null);
-        
-        const newTaskList = {
-            ...data,
-            id: data.id || data.name
-        };
 
-        console.log(newTaskList);
-        dispatch(addTaskList(newTaskList));
+        (
+            data.id
+            ? tasklistService.editTaskList(user.id, data.id, data)
+            : tasklistService.createTaskList(user.id, data)
+        ).then( newTaskList => {
 
-        // If adding new Task List, then continue on the page. Else go back.
-        if (data.id) {
-            handleGoBack();
-        } else {
-            setData(newTaskList);
-            reloadData(newTaskList.id);
-        }
+            console.log(newTaskList);
+            dispatch(addTaskList(newTaskList));
+
+            // If adding new Task List, then continue on the page. Else go back.
+            if (data.id) {
+                handleGoBack();
+            } else {
+                // setData(newTaskList);
+                reloadData(newTaskList.id);
+            }
+
+        })
+        .catch(error => setErrorMessage(error));
+
     }
 
     const handleClickDelete = (e) => {
         e.preventDefault();
 
         if (data.id) {
-            setErrorList(null);
-    
-            dispatch(deleteTaskList(data));
-            handleGoBack();
+
+            const confirmation = confirm(`Are you sure to remove this task list "${data.name}"?`);
+            if (confirmation) {
+                setErrorList(null);
+
+                tasklistService.deleteTaskList(user.id, data.id).then( () => {
+                    dispatch(deleteTaskList(data));
+                    handleGoBack();
+                })
+                .catch(error => setErrorMessage(error));
+            }
         }
     }
 
@@ -116,60 +141,73 @@ const TaskList = () => {
         }
         setErrorTask(null);
 
-        const newTask = {
-            ...taskInitialData,
-            id: taskTitle,
-            title: taskTitle,
-        };
-        const newList = {
-            ...data,
-            tasks: data.tasks.map(item => item)
-        };
-        newList.tasks.push(newTask);
-        newList.id = newList.id || newList.name;
-        setData(newList);
-        setTaskTitle('');
+        taskService.createTask(user.id, data.id, {...taskInitialData, title: taskTitle}).then( newTask => {
+
+            const newList = {
+                ...data,
+                tasks: data.tasks?.map(item => item)
+            };
+            newList.tasks.push(newTask);
+            setData(newList);
+            setTaskTitle('');
+
+            //Save TaskList
+            dispatch(addTaskList(newList));
+            // reloadData(data.id);
+
+        })
+        .catch(error => setErrorMessage(error));
         
-        //Save TaskList
-        dispatch(addTaskList(newList));
-        // reloadData(data.id);
     }
 
     const handleClickUpdateTask = (task, propertyName, newValue) => {
-        const editedTask = {...task};
+        const editedTask = { ...task };
         const levels = {
             high: 'low',
             medium: 'high',
             low: 'medium',
         }
-        editedTask.taskListId = taskList.id;
+        // editedTask.taskListId = taskList.id;
         if (propertyName === 'priority') {
             editedTask['priority'] = levels[editedTask.priority] || 'high';
         } else {
             editedTask[propertyName] = newValue;
         }
-        dispatch(addTask(editedTask));
-        
-        reloadData(data.id);
+
+
+        taskService.editTask(user.id, data.id, task.id, editedTask).then( newTask => {
+            dispatch(addTask(newTask));
+            reloadData(data.id);
+        })
+        .catch(error => setErrorMessage(error));
+
     }
 
     const handleClickDeleteTask = (task) => {
         if (task.id) {
-            const editedTask = {...task};
-            editedTask.taskListId = taskList.id;
-            dispatch(deleteTask(editedTask));
-            
-            reloadData(data.id);
+
+            const confirmation = confirm(`Are you sure to remove this task "${task.title}"?`);
+            if (confirmation) {
+                
+                taskService.deleteTask(user.id, task.tasklist_id, task.id).then( () => {
+                    dispatch(deleteTask(task));
+                    reloadData(task.tasklist_id);
+                })
+                .catch(error => setErrorMessage(error));
+            }
         }
     }
 
     return (
         <form className="flex flex-wrap gap-5">
             <div className="flex flex-1 justify-between">
-                <h1 className="tm-title text-start">
+                <h1 className="tm-title-2 text-start">
                     Task List
                 </h1>
-                <Button text="X" title="Go to Index" className="w-12 text-bold" styleType="transparent" onClick={handleGoBack} />
+                <div className="flex">
+                    <Button title="New Task List" onClick={() => navigate('/task-list')} />
+                    <Button text="X" title="Go to Index" className="w-12 ms-3 text-bold" styleType="transparent" onClick={handleGoBack} />
+                </div>
             </div>
             <Input type="text" name="name" title="List Name" value={data.name} error={errorList} onChange={handleFieldChange} />
 
@@ -182,18 +220,19 @@ const TaskList = () => {
                             value={taskTitle}
                             onChange={handleTaskTitleChange}
                             error={errorTask}
-                            />
+                        />
                         <Button type="submit" title="Add Task" styleType="primary" className="w-24" onClick={handleAddTask} />
                     </form>
 
                     <div className="w-full">
-                        * My Tasks
                         <TaskDetail tasks={data.tasks} status="inProgress" handleClickUpdateTask={handleClickUpdateTask} handleSelectTask={handleSelectTask} handleClickDeleteTask={handleClickDeleteTask} />
                         <TaskDetail tasks={data.tasks} status="todo" handleClickUpdateTask={handleClickUpdateTask} handleSelectTask={handleSelectTask} handleClickDeleteTask={handleClickDeleteTask} />
                         <TaskDetail tasks={data.tasks} status="done" handleClickUpdateTask={handleClickUpdateTask} handleSelectTask={handleSelectTask} handleClickDeleteTask={handleClickDeleteTask} />
                     </div>
                 </>
             }
+
+            { error && <div className="w-full text-sm text-red-500">{error}</div>}
 
             {
                 data.id &&
